refactor(notify): #537 — extract scripts/notify.py to findajob.notifications package
Second M3 extraction. Import-only — no behavior change, no LOC
consolidation. Cross-script duplication (notify() in triage/prep/
interview) preserved for cleanup PR.

Layout (per #538 — 8 commands → 8 per-command modules + shared infra):
- findajob/notifications/cli.py — COMMANDS dispatch + main()
- findajob/notifications/ntfy.py — send(), _persist_notification(),
  db_connect(), recent_log_events(), open_issues(), _p(),
  NOTIFICATION_KINDS, plus `_runtime()` lazy accessor for env-derived
  globals
- findajob/notifications/{daily_stats,apply_reminder,feedback_review,
  health_check,issues_ping,ci_check,scoreboard,send_raw}.py — one
  command per file

Module-load side effects relocated:
- `_env = load_env()` → into a `functools.cache`-d `_runtime()` accessor
  in ntfy.py. First call on a code path that needs it triggers the
  read; importing the modules does not.
- `NTFY_TOPIC = ...` / `NTFY_URL = ...` / `WEB_BASE_URL = ...` → derived
  from `_runtime()` on each call (cheap; cached). `send()` calls
  `_ntfy_url()`; cmd_feedback_review calls `_runtime()["web_base_url"]`.
- `DB_PATH` and `LOG_PATH` stay as module constants (pure f-strings, no
  I/O).

scripts/notify.py shrinks from 984 → 22 LOC. Module sizes (all under
the 300 LOC soft cap):
  scoreboard.py 287, health_check.py 241, ntfy.py 219,
  daily_stats.py 68, apply_reminder.py 61, feedback_review.py 55,
  ci_check.py 52, cli.py 34, send_raw.py 27, issues_ping.py 17.

scoreboard.py preserves the dynamic-import-from-scripts hack for
analyze_feedback.py via sys.path.insert, since analyze_feedback is
itself an M3+ extraction candidate; behaviour is identical.

Tests:
- New tests/test_notify_import_safety.py (5 tests) — locks the import-
  safety invariant: ntfy module load does NOT call load_env(); the
  `_runtime()` cache amortizes to a single load_env() call per process;
  every per-command module imports cleanly; COMMANDS dispatch covers
  the expected 8 keys; NOTIFICATION_KINDS taxonomy unchanged.
- tests/test_notify_persistence.py — drops importlib.util workaround;
  monkeypatches `findajob.notifications.ntfy.DB_PATH` directly.
- tests/test_notify_dejargon.py — drops importlib.util workaround;
  monkeypatches `send` in each per-command module's namespace (where
  `from .ntfy import send` re-binds it). The `test_all_send_titles_are_branded`
  scan now globs the per-command modules instead of one file.
- tests/test_companies_of_interest_consumers.py — read path moved from
  scripts/notify.py to src/findajob/notifications/health_check.py
  (where `is_company_of_interest` is actually called).
- tests/test_crontab_notify_alignment.py — AST-parses
  src/findajob/notifications/cli.py instead of scripts/notify.py;
  handles `AnnAssign` (typed COMMANDS) in addition to `Assign`.

CI green locally: 1854 passed, 1 skipped; ruff check + ruff format
--check both clean repo-wide; mypy clean on the new package (per the
new `feedback_mypy_scope_at_extraction` memory's pre-push checklist).
