+++ src/findajob/notifications/__init__.py
@@ -1,0 +1,6 @@
+"""ntfy push notification suite for findajob.
+
+Subcommand entry-points live in per-command modules; the dispatch table
+and `__main__` guard live in `cli.py`. Import-only extraction from
+`scripts/notify.py` per M3 (#537).
+"""
